Add FileSystemHandle::Remove method
Currently, it is not possible to remove a file or directory given its
handle. You must obtain the handle of the parent directory and call
FileSystemDirectoryHandle::RemoveEntry.

This CL allows for the removal of a file or directory from the local
file system without needing the handle of its parent.

Bug: 1114923
Change-Id: I1db1ef54e6c3a15069a3b8afa100b34f3f912e00