[css-tables] Tests about cells' max-width and excess width distribution
max-width is sometimes honored, sometimes not, when distributing tables'
excess width according to the spec and Firefox. Chrome doesn't honor it
in either of these tests.

I expect that we have some very old tests that assert the opposite of
this new failing test cell-max-width-001; we'll have to figure out which
we want to honor in TablesNG and maybe lobby for a spec change :/

These are tests only, no behavior change.

Bug: 482057
Change-Id: I815fc0c689a2043933983beb9ffcb6c46778adea