Fix wptserve logging when used with wptrunner
Add a LoggerProxy class which just delegates to a logger that writes
the raw log messages to a queue. This LoggerProxy is safe to pass down
to different processes as it only contains pickalable data.

The implementation isn't ideal; we want to reuse the mozlog data
format so that we can directly log the resulting queue entries, but a
lot of the internal state that mozlog has doesn't really make sense
for this use. So we end up having an internal logger that replaces the
normal handlers with a queue and a frontend that only implements the
basic logging functions.