Allow specifying an explicit logger object in wptserve config
wptserve config is intended to be immutable and passed between
multiple server processes. Previously the way the logging worked was
to pass around the name of the logger and create a logger with that
name in each process. But in multi-process scenarios it can also make
sense to conduct all the logging in the parent process (or another
specific process). This happens to be what wptrunner wants. To support
this we introduce a proxy_logger arugment to the ConfigBuilder, which
allows passing in an object that should be used as config.logger in
all processes. This object must therefore be pickle-safe.