Count uncanceled pointermove as "input" for hadRecentIinput
We should not count the layout shift while doing touch dragging or
resizing widgets. While doing touch dragging, we send out pointermove
events, and we do not send pointercancel event. Touch scrolling sends out
pointercancel event to cancel all the active touch pointers and does not
send any more pointermove events when the scroll starts. Therefore, we
add the uncanceled pointermove to the event list that trigger the
exclusion window, which sets hadRecentInput to be true, then we do not
compute the cumulative layout shift score for touch dragging actions.

Bug: 1166952
Change-Id: Ieaaf8a2f5549f078585b706dae34bfbbbd8c501c