Properly compute the trailing space width in RTL text
In the ShapeResult::PositionForOffset we operate with logical and
visual offsets to deal with RTL text.

For RTL, the trailing space position is computed in visual order, so
the value is already the actual trailing space width.

This CL still fails for some cases [1], due to our lack of full support
of UAX#9 L1. We added some support in https://crrev.com/c/2505486 but
we still need to address additional cases.

[1] text-justify-and-trailing-spaces-003.html

Bug: 1187937, 316409
Change-Id: I199ba3056b0c61ec482d6f3c1cb3fe9ff7e0a907