[idlharness.js] Drop validation test created by idl_test()
This was testing the spec and not the implementation.

The IDL in interfaces/*.idl (except *.tentative.idl) is already
validated individually and all together by webidl2.js in the webref
tests which run before @webref/idl is published:
https://github.com/w3c/webref/blob/703e161604ffa6f1ed9619a8786e0aec24b55eba/test/idl/validate.js

Also remove the "extended attributes" validation subtest.

Part of https://github.com/web-platform-tests/wpt/issues/28678.